Split out garbage collector from vm.c/.h
