+++ src/core/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(morpho
     PRIVATE
         compile.c   compile.h
+        gc.c        gc.h
         optimize.c  optimize.h
         vm.c        vm.h
         core.h
